Add function definition, lambda literal for nyaa syntax. Fix handle scope bug(s).

+++ src/nyaa/malloc-object-factory.cc
@@ -107,8 +107,8 @@ MallocObjectFactory::MallocObjectFactory(NyaaCore *core) : core_(DCHECK_NOTNULL(
 }
     
 /*virtual*/ NyFunction *
-MallocObjectFactory::NewFunction(NyString *name, uint8_t n_params, bool vargs, uint32_t n_upvals,
-                                 uint32_t max_stack, NyString *file_name, NyInt32Array *file_info,
+MallocObjectFactory::NewFunction(NyString *name, size_t n_params, bool vargs, size_t n_upvals,
+                                 size_t max_stack, NyString *file_name, NyInt32Array *file_info,
                                  NyByteArray *bcbuf, NyArray *proto_pool, NyArray *const_pool,
                                  bool old) {
     DCHECK_LE(n_params, UINT8_MAX);
