+++ src/nyaa/malloc-object-factory.h
@@ -24,8 +24,8 @@ public:
     virtual NyArray *NewArray(size_t capacity, NyArray *base, bool old) override;
     virtual NyDelegated *NewDelegated(DelegatedKind kind, Address fp, size_t n_upvals,
                                       bool old) override;
-    virtual NyFunction *NewFunction(NyString *name, uint8_t n_params, bool vargs, uint32_t n_upvals,
-                                    uint32_t max_stack, NyString *file_name, NyInt32Array *file_info,
+    virtual NyFunction *NewFunction(NyString *name, size_t n_params, bool vargs, size_t n_upvals,
+                                    size_t max_stack, NyString *file_name, NyInt32Array *file_info,
                                     NyByteArray *bcbuf, NyArray *proto_pool, NyArray *const_pool,
                                     bool old) override;
     virtual NyClosure *NewClosure(NyFunction *proto, bool old) override;
